Moving funciton to fides client to avoid circular imports
